refactor: final audit fixes — types, silent deletes, auth router split
Last-pass review findings:
- Remove all 10 redundant 'as Chunk' casts; the literals were always
  assignable. Deleting them exposed a real string|null gap in
  recover-quiz (blockId used where TS couldn't prove non-null), now
  handled with explicit guards matching endBlock's existing style;
  repair-quiz-parts narrows on the typed chunk instead of a loose cast
- Share one Chunk type via chat/ui-chunks.ts (was defined 3x)
- Admin deleteChatbot/removeDomain now 404 on unknown ids instead of
  returning success on deletes that matched nothing
- Extract useDebouncedValue to hooks/use-debounced-value.ts (was
  duplicated privately in ConversationsTab)
- Split server/routers/auth.ts (434 lines) into directory router:
  status/profile/password/delete-account procedures, verbatim move
- Fix stale 'chat.ts' doc references left from the AI SDK migration
