+++ apps/web/src/hooks/use-debounced-value.ts
@@ -1,0 +1,14 @@
+import { useEffect, useState } from "react";
+
+/**
+ * Returns `value`, delayed by `delayMs`. The timer resets whenever `value`
+ * changes, so rapid updates only settle once the caller stops emitting.
+ */
+export function useDebouncedValue<T>(value: T, delayMs: number): T {
+  const [debounced, setDebounced] = useState(value);
+  useEffect(() => {
+    const t = setTimeout(() => setDebounced(value), delayMs);
+    return () => clearTimeout(t);
+  }, [value, delayMs]);
+  return debounced;
+}
